requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning and AI
scikit-learn>=1.0.0
scipy>=1.7.0

# Data visualization
plotly>=5.0.0
ipywidgets>=7.6.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6.0
flake8>=3.9.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
tqdm>=4.61.0
click>=8.0.0
pyyaml>=5.4.0
python-dateutil>=2.8.0

# Optional: For advanced cryptographic features
cryptography>=3.4.0

# Optional: For web interface integration
flask>=2.0.0
flask-cors>=3.0.0

# Social media API integration
python-dotenv>=0.19.0
requests>=2.25.0